Move `DeprecatedException` to `atopile.errors`
