+++ src/atopile/buildutil.py
@@ -10,8 +10,7 @@ from more_itertools import first
 
 from atopile import layout
 from atopile.config import config
-from atopile.errors import UserException, UserPickError
-from atopile.front_end import DeprecatedException
+from atopile.errors import DeprecatedException, UserException, UserPickError
 from faebryk.core.module import Module
 from faebryk.core.parameter import Parameter
 from faebryk.core.solver.defaultsolver import DefaultSolver
